// src/app/models/watchlistModel.ts
export interface watchlistModel {
  adult: boolean;
  backdrop_path: string;
//   genre_ids: [878, 28, 12];
  id: number;
  original_language: 'en';
  original_title: string;
  overview: string;
//   popularity: 497.324;
  poster_path:string;
  release_date: string;
  title: string;
  video: boolean;
  vote_average: number;
  vote_count: number;


}